Get rid of fixed-precision arithmetic.
It only saves compute time if everything is already in
fixed-precision numbers. Not only is this rarely the case,
dealing with negatives is too confusing.

So there's a performance hit from having to compute the modulo-1
of floats, but overall things are faster since there aren't
extraneous conversions between number representations.